backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
aiosqlite==0.19.0
bcrypt==4.1.2
paramiko==3.3.1
psutil==5.9.6
fabric==3.2.2
cryptography==45.0.2
boto3==1.34.0
python-nmap==0.7.1
netaddr==0.10.1
requests==2.31.0
python-crontab==3.0.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
